feat: move search bar to header
