+++ src/components/Header/index.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react'
+import { useEffect, useRef, useState } from 'react'
 import AppBar from '@mui/material/AppBar'
 import Toolbar from '@mui/material/Toolbar'
 import Typography from '@mui/material/Typography'
@@ -6,15 +6,100 @@ import IconButton from '@mui/material/IconButton'
 import HelpIcon from '@mui/icons-material/Help'
 import SettingsIcon from '@mui/icons-material/Settings'
 import WarningRoundedIcon from '@mui/icons-material/WarningRounded'
-import { Box, Button, Chip, Menu, MenuItem, Stack, Tooltip } from '@mui/material'
+import { Box, Button, Chip, Menu, MenuItem, Stack, TextField, Tooltip } from '@mui/material'
 import LoadingButton from '@mui/lab/LoadingButton'
-import { useAppStore } from '../../stores'
+import { useAppStore, useNodeStore } from '../../stores'
 import { signOut } from '../../lens/auth'
 import { JWT_ADDRESS_KEY } from '../../constants'
 import { useAccount } from 'wagmi'
 import { ProfilePicture } from '../Shared/ProfilePicture'
 import { useQuery } from 'urql'
 import { graphql } from '../../lens/schema'
+import { useSnackbar } from 'notistack'
+import { getProfileNode } from '../../lens/profile'
+import { TooManyFollowingException } from '../../errors'
+
+const SearchBar = () => {
+  const addNodes = useNodeStore((state) => state.addNodes)
+  const selectNode = useAppStore((state) => state.selectNode)
+  const isQuerying = useAppStore((state) => state.isQuerying)
+  const setIsQuerying = useAppStore((state) => state.setIsQuerying)
+  const handleInputRef = useRef<HTMLInputElement>()
+  const { enqueueSnackbar } = useSnackbar()
+
+  const handleAdd = async () => {
+    if (isQuerying) return
+    try {
+      setIsQuerying(true)
+      const handle = handleInputRef?.current?.value
+      if (!handle) {
+        enqueueSnackbar('Please enter a non-empty handle', { variant: 'error' })
+        return
+      }
+      try {
+        const { profile } = await getProfileNode(handle)
+        addNodes([profile])
+        selectNode(profile.id)
+        return
+      } catch (error) {
+        if (error instanceof TooManyFollowingException) {
+          enqueueSnackbar(`${handle} is following too many profiles`, { variant: 'error' })
+          return
+        }
+
+        if (!handle.endsWith('.lens')) {
+          // try again with '.lens' appended
+          try {
+            const handleWithLens = handle.concat('.lens')
+            const { profile } = await getProfileNode(handleWithLens)
+            addNodes([profile])
+            selectNode(profile.id)
+            return
+          } catch (error) {
+            if (error instanceof TooManyFollowingException) {
+              enqueueSnackbar(`${handle} is following too many profiles`, { variant: 'error' })
+              return
+            }
+          }
+        }
+        enqueueSnackbar(`Handle not found: ${handle}`, { variant: 'error' })
+      }
+    } finally {
+      setIsQuerying(false)
+    }
+  }
+
+  const handleKeyPress = (evt: React.KeyboardEvent<HTMLDivElement>) => {
+    if (evt.key === 'Enter') {
+      evt.preventDefault()
+      handleAdd()
+    }
+  }
+
+  return (
+    <>
+      <TextField
+        onKeyDown={handleKeyPress}
+        inputRef={handleInputRef}
+        name='handle'
+        required
+        fullWidth
+        id='handle'
+        label='Lens Handle'
+        defaultValue='cultivator'
+        autoFocus
+        disabled={isQuerying}
+      />
+      <LoadingButton
+        loading={isQuerying}
+        onClick={handleAdd}
+        variant='contained'
+      >
+        Add
+      </LoadingButton>
+    </>
+  )
+}
 
 const SettingsButton = () => {
   const showSettings = useAppStore((state) => state.showSettings)
@@ -169,7 +254,7 @@ const Header = () => {
           Cultivator
         </Typography>
         <Chip icon={<WarningRoundedIcon />} onClick={() => setShowBeta(true)} label='Beta' color='warning' sx={{ ml: 1.5, p: 0.5 }} />
-
+        <SearchBar />
         <Box sx={{ flexGrow: 1 }} />
 
         <Stack direction='row' spacing={1.3}>
